+++ docs/src/examples/components/Dropdown/Types/index.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+import ComponentExample from 'docs/src/components/ComponentDoc/ComponentExample'
+import ExampleSection from 'docs/src/components/ComponentDoc/ExampleSection'
+
+const Types = () => (
+  <ExampleSection title="Types">
+    <ComponentExample
+      title="Multiple Search"
+      description="A dropdown with multiple selection and search."
+      examplePath="components/Dropdown/Types/DropdownExampleMultipleSearch.shorthand"
+    />
+  </ExampleSection>
+)
+
+export default Types