+++ src/themes/teams/staticStyles/globalStyles.ts
@@ -8,6 +8,7 @@ const globalStyles: StaticStyleFunction = siteVars => ({
     padding: siteVars.bodyPadding,
     margin: siteVars.bodyMargin,
     fontFamily: siteVars.bodyFontFamily,
+    fontSize: siteVars.bodyFontSize,
     lineHeight: siteVars.bodyLineHeight,
   },
 })