+++ src/themes/types.ts
@@ -284,6 +284,9 @@ export interface ThemeComponentStylesInput {
   ChatItem?: ComponentSlotStylesInput
   ChatMessage?: ComponentSlotStylesInput
   Divider?: ComponentSlotStylesInput
+  Dropdown?: ComponentSlotStylesInput
+  DropdownItem?: ComponentSlotStylesInput
+  DropdownSearchInput?: ComponentSlotStylesInput
   Form?: ComponentSlotStylesInput
   FormField?: ComponentSlotStylesInput
   Grid?: ComponentSlotStylesInput
@@ -322,6 +325,9 @@ export interface ThemeComponentStylesPrepared {
   ChatItem?: ComponentSlotStylesPrepared
   ChatMessage?: ComponentSlotStylesPrepared
   Divider?: ComponentSlotStylesPrepared
+  Dropdown?: ComponentSlotStylesPrepared
+  DropdownItem?: ComponentSlotStylesPrepared
+  DropdownSearchInput?: ComponentSlotStylesPrepared
   Form?: ComponentSlotStylesPrepared
   FormField?: ComponentSlotStylesPrepared
   Grid?: ComponentSlotStylesPrepared
@@ -360,6 +366,7 @@ export interface ThemeComponentVariablesInput {
   ChatItem?: ComponentVariablesInput
   ChatMessage?: ComponentVariablesInput
   Divider?: ComponentVariablesInput
+  Dropdown?: ComponentVariablesInput
   Form?: ComponentVariablesInput
   FormField?: ComponentVariablesInput
   Grid?: ComponentVariablesInput
@@ -398,6 +405,7 @@ export interface ThemeComponentVariablesPrepared {
   ChatItem?: ComponentVariablesPrepared
   ChatMessage?: ComponentVariablesPrepared
   Divider?: ComponentVariablesPrepared
+  Dropdown?: ComponentVariablesPrepared
   Form?: ComponentVariablesPrepared
   FormField?: ComponentVariablesPrepared
   Grid?: ComponentVariablesPrepared